requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
datasets==2.18.0
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.1
FlagEmbedding==1.2.5
Flask==3.0.2
Flask-GraphQL==2.0.1
Flask-PyMongo==2.3.0
frozenlist==1.4.1
fsspec==2024.2.0
graphene==3.3
graphql-core==2.3.2
graphql-relay==3.2.0
graphql-server-core==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.1
numpy==1.24.4
openai==1.13.3
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
pillow==10.2.0
promise==2.3
psutil==5.9.8
pyactiveresource==2.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
Rx==1.6.3
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.5.1
ShopifyAPI==12.4.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
